Rename LoginForm component to AuthForm
